Lab_05/requirements.txt
requests~=2.25.1
Flask~=1.1.2
hazelcast-python-client